Bernoulli Animations
Added animation of transformation as well as labels for associated Bernoulli process.

I realized x = 1 is a problem point for how the transformation is defined. Except for particularly "nice" values of x (of form of 1/2^n) there shouldn't be an issue with getting "stuck" at 1. I think I need higher precision arithmetic to stop the transformation from rounding to these "nice" x values.